Adding ipython files.
This code developed is an example of SQL-Python automation. The code perfectly shows how SQL queries can be clubbed with Python and daily tasks can be automated. A sample SQL table is created in the form of a dataframe and how it is queried. The code is embedded with email script which sends email updates as soon as the query completes based on a scheduled time.